+++ src/components/configurable-items/index.d.ts
@@ -1,2 +1,0 @@
-export { default as ConfigurableItems } from "./configurable-items";
-export { default as ConfigurableItemRow } from "./configurable-item-row/configurable-item-row";